[FRONT] Separate "Homepage" and "Dashboard"
