+++ nontclient/src/Components/Dashboard/NontOwnerDashboard.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const NontOwnerDashboard = () => {
+  return ( 
+    <h1>Nont Owner Dashboard</h1>
+  );
+}
+ 
+export default NontOwnerDashboard;
