+++ nontclient/src/Components/Dashboard/NontSitterDashboard.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const NontSitterDashboard = () => {
+  return ( 
+    <h1>Nont Sitter Dashboard</h1>
+  );
+}
+ 
+export default NontSitterDashboard;
